Split files, fix bug
